chore: refactoring notifier_bot module, introducing NotifBotAppBuilder
